+++ gateway/src/apicast/http_proxy.lua
@@ -55,6 +55,7 @@ local function forward_https_request(proxy_uri, uri, proxy_opts)
     local content_type = ngx_req_get_headers()["Content-Type"]
     local content_type_is_urlencoded = content_type and content_type:lower() == "application/x-www-form-urlencoded"
     local raw = false
+    local request_start = ngx.now()
 
     if http_methods_with_body[req_method] then
 
@@ -165,6 +166,9 @@ local function forward_https_request(proxy_uri, uri, proxy_opts)
     res, err = httpc:request(request)
 
     if res then
+        ngx.ctx.proxy_upstream_status = res.status
+        ngx.ctx.proxy_upstream_response_time = ngx.now() - request_start
+
         if opts.request_unbuffered and raw then
             err = send_response(sock, res, DEFAULT_CHUNKSIZE)
             if err then